backend/assignTicket/requirements.txt
functions-framework==3.*
google-cloud-pubsub
flask
requests